odibi_core/cloud/kafka_adapter.py
"""
Kafka Adapter - Simulation Stub (Phase 7)

This is a simulation stub for future Kafka streaming integration.
Set simulate=True to test Kafka-based pipelines without actual Kafka cluster.

Future implementation will include:
- Kafka producer/consumer for streaming data
- Topic management and offset tracking
- Integration with Confluent Cloud
- AVRO/JSON schema support
- Consumer group management
"""

import logging
from typing import Any, List, Optional, Dict
import pandas as pd

from .cloud_adapter import CloudAdapterBase

logger = logging.getLogger(__name__)


class KafkaAdapter(CloudAdapterBase):
    """
    Kafka adapter (simulation stub).

    This adapter provides a simulation interface for Kafka streaming operations.
    For production use, implement actual Kafka integration using confluent-kafka or kafka-python.

    Usage:
        adapter = KafkaAdapter(
            bootstrap_servers="localhost:9092",
            topic="my-topic",
            simulate=True
        )
        adapter.connect()

        # Consume messages (simulated)
        messages = adapter.read(max_messages=100)

        # Produce messages (simulated)
        adapter.write(df, topic="output-topic")
    """

    def __init__(
        self,
        bootstrap_servers: Optional[str] = None,
        topic: Optional[str] = None,
        group_id: Optional[str] = None,
        security_protocol: str = "PLAINTEXT",
        sasl_mechanism: Optional[str] = None,
        sasl_username: Optional[str] = None,
        sasl_password: Optional[str] = None,
        simulate: bool = True,  # Default to simulation
        **kwargs,
    ):
        """
        Initialize Kafka adapter.

        Args:
            bootstrap_servers: Kafka broker addresses (e.g., "localhost:9092")
            topic: Default Kafka topic
            group_id: Consumer group ID
            security_protocol: Security protocol (PLAINTEXT, SASL_SSL, etc.)
            sasl_mechanism: SASL mechanism (PLAIN, SCRAM-SHA-256, etc.)
            sasl_username: SASL username
            sasl_password: SASL password
            simulate: Must be True for Phase 7 (no real implementation yet)
            **kwargs: Additional Kafka config
        """
        super().__init__(simulate=simulate, **kwargs)

        self.bootstrap_servers = bootstrap_servers
        self.topic = topic
        self.group_id = group_id
        self.security_protocol = security_protocol
        self.sasl_mechanism = sasl_mechanism
        self.sasl_username = sasl_username
        self.sasl_password = sasl_password

        if not simulate:
            logger.warning(
                "KafkaAdapter is a simulation stub in Phase 7. "
                "Set simulate=True or implement confluent-kafka integration for production use."
            )
            self.simulate = True  # Force simulation mode

    def connect(self) -> bool:
        """Establish connection to Kafka (simulated)"""
        logger.info(
            f"[SIMULATE] Connected to Kafka: {self.bootstrap_servers}, "
            f"topic: {self.topic}, group: {self.group_id}"
        )
        self.connected = True
        return True

    def read(
        self,
        topic: Optional[str] = None,
        max_messages: int = 100,
        timeout: float = 1.0,
        **kwargs,
    ) -> pd.DataFrame:
        """
        Consume messages from Kafka topic (simulated).

        Args:
            topic: Kafka topic to consume from (uses default if None)
            max_messages: Maximum messages to consume
            timeout: Consumer timeout in seconds
            **kwargs: Additional consumer options

        Returns:
            Simulated DataFrame with message data
        """
        topic = topic or self.topic
        logger.info(
            f"[SIMULATE] Consuming up to {max_messages} messages from Kafka topic: {topic}"
        )

        # Return simulated messages
        import random

        num_messages = random.randint(1, max_messages)

        return pd.DataFrame(
            {
                "topic": [topic] * num_messages,
                "partition": [random.randint(0, 3) for _ in range(num_messages)],
                "offset": list(range(num_messages)),
                "key": [f"key_{i}" for i in range(num_messages)],
                "value": [f"message_{i}" for i in range(num_messages)],
                "timestamp": pd.date_range(
                    "2025-01-01", periods=num_messages, freq="1s"
                ),
            }
        )

    def write(
        self,
        data: pd.DataFrame,
        topic: Optional[str] = None,
        key_column: Optional[str] = None,
        value_column: Optional[str] = None,
        **kwargs,
    ) -> bool:
        """
        Produce messages to Kafka topic (simulated).

        Args:
            data: DataFrame to write
            topic: Kafka topic to produce to (uses default if None)
            key_column: Column to use as message key
            value_column: Column to use as message value (uses entire row as JSON if None)
            **kwargs: Additional producer options

        Returns:
            True if write successful
        """
        topic = topic or self.topic
        logger.info(
            f"[SIMULATE] Producing {len(data)} messages to Kafka topic: {topic}"
        )

        if key_column:
            logger.info(f"[SIMULATE] Using {key_column} as message key")
        if value_column:
            logger.info(f"[SIMULATE] Using {value_column} as message value")
        else:
            logger.info("[SIMULATE] Using entire row as JSON value")

        return True

    def exists(self, topic: str) -> bool:
        """Check if Kafka topic exists (simulated)"""
        logger.info(f"[SIMULATE] Checking Kafka topic existence: {topic}")
        return True

    def list(self, pattern: Optional[str] = None) -> List[str]:
        """List Kafka topics (simulated)"""
        logger.info("[SIMULATE] Listing Kafka topics")
        return ["topic1", "topic2", "energy-sensors", "processed-data"]

    def delete(self, topic: str) -> bool:
        """Delete Kafka topic (simulated)"""
        logger.info(f"[SIMULATE] Deleting Kafka topic: {topic}")
        return True

    def create_topic(
        self, topic: str, num_partitions: int = 3, replication_factor: int = 1
    ):
        """Create Kafka topic (simulated)"""
        logger.info(
            f"[SIMULATE] Creating Kafka topic: {topic} "
            f"(partitions={num_partitions}, replication={replication_factor})"
        )
        return True

    def get_offsets(self, topic: Optional[str] = None) -> Dict[int, int]:
        """Get current consumer offsets (simulated)"""
        topic = topic or self.topic
        logger.info(f"[SIMULATE] Getting consumer offsets for topic: {topic}")

        # Return simulated offsets
        return {0: 1234, 1: 2345, 2: 3456}


# TODO: Implement real Kafka integration using confluent-kafka
# Example implementation outline:
"""
from confluent_kafka import Producer, Consumer, KafkaError

class KafkaAdapterProduction(CloudAdapterBase):
    def __init__(self, bootstrap_servers, topic, group_id="odibi-consumer", **kwargs):
        super().__init__(simulate=False, **kwargs)
        self.bootstrap_servers = bootstrap_servers
        self.topic = topic
        self.group_id = group_id
        self.producer = None
        self.consumer = None
    
    def connect(self):
        producer_config = {'bootstrap.servers': self.bootstrap_servers}
        consumer_config = {
            'bootstrap.servers': self.bootstrap_servers,
            'group.id': self.group_id,
            'auto.offset.reset': 'earliest'
        }
        self.producer = Producer(producer_config)
        self.consumer = Consumer(consumer_config)
        self.consumer.subscribe([self.topic])
        return True
    
    def read(self, max_messages=100, timeout=1.0):
        messages = []
        for _ in range(max_messages):
            msg = self.consumer.poll(timeout)
            if msg is None:
                break
            if msg.error():
                continue
            messages.append({
                'key': msg.key(),
                'value': msg.value(),
                'partition': msg.partition(),
                'offset': msg.offset()
            })
        return pd.DataFrame(messages)
    
    def write(self, data, topic=None, key_column=None, value_column=None):
        topic = topic or self.topic
        for _, row in data.iterrows():
            key = row[key_column] if key_column else None
            value = row.to_json() if not value_column else row[value_column]
            self.producer.produce(topic, key=key, value=value)
        self.producer.flush()
        return True
"""
