requirements-dev.txt
# ODIBI CORE v1.0 - Development Dependencies
# Install with: pip install -r requirements-dev.txt

# Include base requirements
-r requirements.txt

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
black>=23.0.0
mypy>=1.0.0
pandas-stubs>=1.5.0

# Interactive development
ipython>=8.0.0
jupyter>=1.0.0
